Add copy_to_st2kv actions
These actions read values from Vault and then save them
in the ST2 datastore. They should be retrieved in workflows via the `st2kv` function.

StackStorm doesn't have a way to mark action output as "secret" so that it doesn't show up in CLI or St2Web output.
This works around that by using the datastore as an encrypted pipeline.